finished order and cart

+++ src/main/java/gov/iti/jets/dto/CustomerDto.java
@@ -18,7 +18,7 @@ public class CustomerDto extends BaseDTO {
 
     private  Date birthday;
     private  String email;
-    private  Integer creditLimit;
+    private  Double creditLimit;
     private Boolean isAdmin;
